requirements.txt
numpy==1.26.2
pandas==2.1.3
Pillow==10.1.0
Pillow==10.1.0
streamlit==1.29.0
tifffile==2023.9.26
torch==2.1.1
torchvision==0.16.1
